64-bit warnings cleanup
Fixed almost all warnings when compiling for 64-bit, and fixed Lua
invocations that were using lua_Number when they should have been using
lua_Integer. Lua still crashes in 64-bit builds, but it appears to get a
bit farther than it did before.

+++ include/Inventory.h
@@ -30,7 +30,7 @@ public:
 
   /// Gets the current count of this inventory.
   /// @return The number of items in the inventory.
-  unsigned int count();
+  size_t count();
 
   /// Gets a beginning iterator to the things map.
   ThingMap::iterator begin();
